begin refactor out of iview components of wlutz arc
