+++ lib/pymedphys/_experimental/streamlit/utilities/iview/_utilities.py
@@ -14,40 +14,12 @@
 
 import pathlib
 
-from pymedphys._imports import natsort
 from pymedphys._imports import numpy as np
 
 from pymedphys._streamlit.utilities import config as _config
 from pymedphys._streamlit.utilities import misc
 
 from . import _dbf
-
-
-def iterate_over_columns(dataframe, data, columns, callbacks, previous=tuple()):
-    column = columns[0]
-    callback = callbacks[0]
-
-    sorted_items = natsort.natsorted(dataframe[column].unique())
-    for item in sorted_items:
-        filtered_dataframe = filter_by(dataframe, column, item)
-        if len(filtered_dataframe) == 0:
-            continue
-
-        args = previous + (item,)
-
-        if callback is not None:
-            callback(filtered_dataframe, data, *args)
-
-        if len(columns) > 1:
-            iterate_over_columns(
-                filtered_dataframe, data, columns[1:], callbacks[1:], previous=args
-            )
-
-
-def filter_by(dataframe, column, value):
-    filtered = dataframe.loc[dataframe[column] == value]
-
-    return filtered
 
 
 def filepath_to_filename(path):
